backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.0
python-multipart>=0.0.9
greenlet>=3.0.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bleach>=6.0.0

# Monitoring
python-json-logger>=2.0.7
